requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-jazzmin==2.6.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
packaging==23.2
Pillow==10.1.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
replicate==0.22.0
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
